% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{condition handling}
\usage{
f_operator(f)
}
\arguments{
\item{f}{(a function)}
}
\value{
a function
}
\description{
condition handling
}
\examples{
f<-log_2
log_3 <- f_operator(f)
}
