% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_2.R
\name{log_2}
\alias{log_2}
\title{my version of function log}
\usage{
log_2(x)
}
\arguments{
\item{x}{}
}
\value{
number or error
}
\description{
my version of function log
}
\examples{
x=2
log_2(x)
}
\seealso{
x=-1
log_2(x)
}
