% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_2.R
\name{sqrt_2}
\alias{sqrt_2}
\title{my version of log function}
\usage{
sqrt_2(x)
}
\arguments{
\item{x}{}
}
\value{
number or error
}
\description{
my version of log function
}
\examples{
x=2
sqrt(x)
}
\seealso{
x=-1
sqrt(x)
}
